requirements.txt
plotly==5.14.1
python-dotenv==1.0.0
requests==2.28.2
